pyproject.toml

[project]
version = '0.0.0'
name = 'Woods'

requires-python = '>=3.10'

dependencies = [
    'freecad-stubs>=1.0.21'
]